# backend/internal/db/sqlc/storage/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package storage

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Announcement struct {
	ID        uuid.UUID
	Message   string
	CreatedAt pgtype.Timestamptz
	CreatedBy uuid.UUID
	Active    bool
}

type AnnouncementDismissal struct {
	AnnouncementID uuid.UUID
	UserID         uuid.UUID
	DismissedAt    pgtype.Timestamptz
}

type Balance struct {
	UserID    uuid.UUID
	Amount    int32
	UpdatedAt pgtype.Timestamptz
}

type DailyReward struct {
	UserID        uuid.UUID
	LastClaimedAt pgtype.Timestamptz
	StreakDays    int32
}

type Hug struct {
	ID         uuid.UUID
	GiverID    uuid.UUID
	ReceiverID uuid.UUID
	CreatedAt  pgtype.Timestamptz
	Status     string
	AcceptedAt pgtype.Timestamptz
	HugType    string
	Comment    pgtype.Text
	StreakTier string
}

type HugCooldown struct {
	UserAID              uuid.UUID
	UserBID              uuid.UUID
	LastHugAt            pgtype.Timestamptz
	CooldownSeconds      int32
	DeclineCooldownUntil pgtype.Timestamptz
}

type PairIntimacy struct {
	UserAID     uuid.UUID
	UserBID     uuid.UUID
	RawScore    int32
	LastHugAt   pgtype.Timestamptz
	LastDecayAt pgtype.Timestamptz
}

type PairStreak struct {
	UserAID        uuid.UUID
	UserBID        uuid.UUID
	CurrentStreak  int32
	BestStreak     int32
	LastStreakDate pgtype.Date
	AHuggedToday   bool
	BHuggedToday   bool
	TodayDate      pgtype.Date
}

type RefreshToken struct {
	Jti       string
	UserID    uuid.UUID
	ExpiresAt pgtype.Timestamptz
	RevokedAt pgtype.Timestamptz
	CreatedAt pgtype.Timestamptz
}

type SudokuCaptcha struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Puzzle    []byte
	Solution  []byte
	Errors    int32
	Passed    bool
	ExpiresAt pgtype.Timestamptz
	CreatedAt pgtype.Timestamptz
}

type User struct {
	ID                  uuid.UUID
	Username            string
	Password            string
	Role                string
	Gender              pgtype.Text
	BannedAt            pgtype.Timestamptz
	HugSlots            int32
	CreatedAt           pgtype.Timestamptz
	DisplayName         pgtype.Text
	TelegramID          pgtype.Int8
	Tag                 pgtype.Text
	SpecialTag          pgtype.Text
	RequiresSudoku      bool
	SudokuCooldownUntil pgtype.Timestamptz
}

type UserBlock struct {
	BlockerID uuid.UUID
	BlockedID uuid.UUID
	CreatedAt pgtype.Timestamptz
}